+++ man-roxygen/paramset_torchlearner.R
@@ -18,10 +18,6 @@
 #'   divide the available cores among the workers instead of setting this to the number of cores.
 #' * `num_interop_threads` :: `integer(1)`\cr
 #'   The number of threads for interop parallelization (if `device` is `"cpu"`).
-#'   This value is initialized to 1.
-#'   Note that this can only be set **once** per session, so training a learner once already sets this.
-#'   Because the value is initialized, a deviating value only takes effect if it is set on the first
-#'   learner that is trained in a session. You can work around this via encapsulation, see [`mlr3::Learner`].
 #' * `seed` :: `integer(1)` or `"random"` or `NULL`\cr
 #'   The torch seed that is used during training and prediction.
 #'   This value is initialized to `"random"`, which means that a random seed will be sampled at the beginning of the